requirements.txt
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-speech==1.37.0
azure-common==1.1.28
azure-core==1.30.2
azure-search-documents==11.4.0
Flask==3.0.3
Flask-Cors==4.0.1
gremlinpython==3.7.0
openai==1.33.0
pandas==2.2.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0